requirements.txt
flask==2.3.3
Flask-Cors==4.0.0
Flask-Login==0.6.3
cryptocode==0.1
cloudinary==1.38.0
jwt==1.3.1
gunicorn==20.1.0
oauthlib==3.2.2
Werkzeug==2.3.7
bcrypt==4.1.1
pyscopg2==66.0.2
psycopg2_binary==2.9.5
requests==2.31.0
 
